requirements.txt
ipykernel
streamlit
langchain
langchain-community
langchain-chroma
langchain-core
langchain-groq
langchain-huggingface
python-dotenv